Move electron type augmentation to separate dts file
